Added page on, and conda environment for, Augustus
Added a program guide for Augustus.
A new conda environment was added for Augustus as it conflicted
with programs already in the "bioinfo-notebook" environment.
This new conda environment can be installed using "augustus.txt"
or "augustus.yml" in the "envs/" directory.
Commands to verify that this new conda environment works using
continuous integration were added to ".travis.yml".