Added input argument to genome annotation script
Added an input argument to genome_annotation_SwissProt_CDS.sh, to allow
an input FASTA nucleotide file to be annotated. Also added a 'demo'
argument to allow this script to run with demonstration data. Edited the
help text for this script and snp_calling.sh to fit the same style.